+++ src/python/grpcio/grpc_core_dependencies.py
@@ -67,6 +67,8 @@ CORE_SOURCE_FILES = [
     'src/core/ext/filters/client_channel/resolver/xds/xds_resolver.cc',
     'src/core/ext/filters/client_channel/resolver_registry.cc',
     'src/core/ext/filters/client_channel/resolver_result_parsing.cc',
+    'src/core/ext/filters/client_channel/retry_filter.cc',
+    'src/core/ext/filters/client_channel/retry_service_config.cc',
     'src/core/ext/filters/client_channel/retry_throttle.cc',
     'src/core/ext/filters/client_channel/server_address.cc',
     'src/core/ext/filters/client_channel/service_config.cc',
